pyproject.toml
[tool.poetry]
name = "test-fabrika-v"
version = "0.1.0"
description = ""
authors = ["polenom <vitalimitsevich@yandex.by>"]
readme = "README.md"
packages = [{include = "test_fabrika_v"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
celery = "^5.2.7"
django-celery-results = "^2.4.0"
redis = "^4.4.0"
django-redis = "^5.2.0"
django-celery-beat = "^2.4.0"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.2.0"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
